+++ react/__mocks__/vtex.delivery-promise-components/DeliveryPromiseContext.js
@@ -1,2 +1,2 @@
-export const useDeliveryPromiseState = () => ({})
-export const useDeliveryPromiseDispatch = () => () => {}
+export const useDeliveryPromiseState = jest.fn(() => ({}))
+export const useDeliveryPromiseDispatch = jest.fn(() => () => {})